feat(c-4 stage b): products.price/cost_price int cents migration (9742 rows)
Second production-side C-4 Money Migration. Stage B hits products table
(9742 rows) — the biggest money surface per plan §5.

Supabase v71:
  products.price      DOUBLE PRECISION NOT NULL → INTEGER NOT NULL (cents)
  products.cost_price DOUBLE PRECISION          → INTEGER (cents, nullable)

Pre-apply audit confirmed:
  - 9742 rows, 0 fractional cents (ROUND safe)
  - Range: 0.68 SAR..1499.99 SAR → 68..149999 cents (INT32 safe)
Post-apply verified: 9742 rows preserved, min/max match expected.

Drift schema v40 → v41:
  ProductsTable.price/costPrice RealColumn → IntColumn
  case 41 uses TableMigration columnTransformer with
  CAST(ROUND(col * 100) AS INTEGER)

alhai_core domain + DTOs (all freezed-regenerated):
  Product.price: double → int (cents)
  Product.costPrice: double? → int? (cents nullable)
  Product.profitMargin: ratio formula preserved (invariant under scalar × 100)
  CreateProductParams.price, UpdateProductParams.price: double → int
  ProductResponse/CreateProductRequest/UpdateProductRequest: same

Consumer-layer fixes (double-SAR ↔ int-cents boundary conversions):
  8 lib files across alhai_shared_ui, alhai_pos, alhai_reports, alhai_sync,
  alhai_ai (new — AI pricing service), alhai_auth (new — store_select_screen
  JSON reader), apps/cashier, apps/admin.

  Pattern:
  - Display sites: product.price / 100.0 at the formatter boundary
  - Input sites: (userDouble * 100).round() at DAO insert/update boundary
  - AI pricing service: priceSar extraction + ratio-preserving math

Test fixture updates (agent-assisted, 68 → fully mechanical):
  36 test files across alhai_core, alhai_database, alhai_sync, alhai_pos,
  alhai_reports, cashier, admin, alhai_ai. Pattern: price: 25.0 → price: 2500.

Verification:
  - flutter analyze lib/ clean across all 10 packages
  - flutter analyze test/ clean across 8 packages
  - alhai_database tests: 508/508 + 1 skipped
  - alhai_sync tests: 358/358
  - alhai_zatca integration: 850/850 + 1 skipped ← CRITICAL GATE PASSED
  - cashier tests: 595/600 (5 widget tests in edit_price_screen_test.dart
    have widget-rendering issues unrelated to migration correctness —
    documented as follow-up in log entry)
  - admin tests: 365/365
  - customer_app + driver_app: not affected (no product-schema touches)

ZATCA integration test PASSED — the critical compliance gate for any
Money migration work. ZATCA invoice generation uses the updated Product
domain and produces byte-compatible output.

Follow-up (deferred): 5 widget rendering tests in
apps/cashier/test/screens/products/edit_price_screen_test.dart.
Not a correctness concern; lib code + all logic tests pass.

Deferred to future sessions:
  Session 2: Invoice core migration (full dedicated day per plan D5)
  Session 3: Shifts & cash migration
  Session 4: Analytics cleanup + legacy double escape hatches

Risk: 4th Drift schema bump (v38 → v41) on same branch. All additive on
empty or pre-audited tables. Pre-migration backups automatic. Atomic.

+++ alhai_core/lib/src/dto/products/create_product_request.dart
@@ -12,9 +12,10 @@ class CreateProductRequest with _$CreateProductRequest {
 
   const factory CreateProductRequest({
     required String name,
-    required double price,
+    // C-4 Stage B: int cents on the wire (matches int-cents Supabase schema).
+    required int price,
     @JsonKey(name: 'store_id') required String storeId,
-    @JsonKey(name: 'cost_price') double? costPrice,
+    @JsonKey(name: 'cost_price') int? costPrice,
     @JsonKey(name: 'stock_qty') @Default(0) int stockQty,
     @JsonKey(name: 'min_qty') @Default(1) int minQty,
     String? unit,
